Unimplement TypeNum for *mut ffi::PyObject
